move speech and serial function out of main.py - make them easier to import as libraries
